Move order related methods from `Dispatcher` to `OrderDispatcher`
